+++ laktory/__init__.py
@@ -16,8 +16,4 @@ import laktory.yaml
 
 from ._logger import get_logger
 from ._settings import Settings
-from .cli.app import app
-from .datetime import unix_timestamp
-from .datetime import utc_datetime
-from .dispatcher.dispatcher import Dispatcher
 from .version import show_version_info