+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Release History
 
-## [0.6.1] - Unreleased
+## [0.6.2] - Unreleased
+### Added
+* n/a
+### Fixed
+* n/a
+### Updated
+* n/a
+### Breaking changes
+* n/a
+
+## [0.6.1] - 2025-01-20
 ### Updated
 * Reduced import time
 * Optimized tests to load spark only when required